+++ Samples~/PlayableDemo/DemoHealthBar.cs
@@ -2,7 +2,7 @@
 using Likeon.GAS;
 using UnityEngine;
 
-namespace GASDemo
+namespace Likeon.GAS.Sample.PlayableDemo
 {
     public class DemoHealthBar : MonoBehaviour
     {