+++ Samples~/PlayableDemo/DemoRangedAbility.cs
@@ -3,7 +3,7 @@
 using Likeon.GAS;
 using UnityEngine;
 
-namespace GASDemo
+namespace Likeon.GAS.Sample.PlayableDemo
 {
     public class DemoRangedAbility : GameplayAbility
     {